@yzhang1991 2019-12-11
influxdata/flux#1992 introduced an `EnableExperimentalTracing` function that is included in Flux v0.57.0 #16186. I will wait for it to be merged then update this PR to introduce the environment variable switch